players usando dados a partir do txt

+++ ProjectVikins/Assets/Script/DAL/Player.cs
@@ -11,8 +11,8 @@ namespace Assets.Script.DAL
         public PossibleMoviment LastMoviment { get; set; }
         public bool IsBeingControllable { get; set; }
         public PlayerModes PlayerMode { get; set; }
-        public int InitialX { get; set; }
-        public int InitialY { get; set; }
+        public float InitialX { get; set; }
+        public float InitialY { get; set; }
 
         public Player()
         {
